Chnaged original code with axios
