+++ scripts/emptyFolder.js
@@ -2,7 +2,7 @@ import fs from 'fs-extra';
 
 function emptyDist() {
   return new Promise((resolve, reject) => {
-    fs.emptyDir('../../dist', (err) => {
+    fs.emptyDir('../dist', (err) => {
       if (err) {
         console.error(err);
         reject();
