+++ src/index.js
@@ -1,4 +1,10 @@
-import parser from './parser';
+import GetInstaData from './getInstaData';
 
-parser()
-export default parser
+// parser().then((res) => {
+//   console.log(`all tasks are done! ${res}`);
+// });
+GetInstaData.getTopPosts('트와이스')
+  .then((res) => {
+    console.log(res);
+    console.log('All tasks are done!');
+  });
